+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Add a `PathMap` class that uses path equality for its keys.
 
+* Add a `PathSet` class that uses path equality for its contents.
+
 ## 1.5.1
 
 * Fix a number of bugs that occurred when the current working directory was `/`